+++ MatrixSDK/Utils/MXTools.m
@@ -809,6 +809,7 @@ static NSMutableDictionary *fileExtensionByContentType = nil;
 #pragma mark - Video processing
 
 + (void)convertVideoToMP4:(NSURL*)videoLocalURL
+       withTargetFileSize:(NSInteger)targetFileSize
                   success:(void(^)(NSURL *videoLocalURL, NSString *mimetype, CGSize size, double durationInMs))success
                   failure:(void(^)(void))failure
 {
@@ -824,11 +825,17 @@ static NSMutableDictionary *fileExtensionByContentType = nil;
     NSString *cacheRoot = [paths objectAtIndex:0];
     outputVideoLocalURL = [NSURL fileURLWithPath:[cacheRoot stringByAppendingPathComponent:outputFileName]];
     
-    // Convert video container to mp4
-    // Use medium quality to save bandwidth
+    // Convert video container to mp4 using preset from MXSDKOptions.
     AVURLAsset* videoAsset = [AVURLAsset URLAssetWithURL:videoLocalURL options:nil];
-    AVAssetExportSession *exportSession = [AVAssetExportSession exportSessionWithAsset:videoAsset presetName:AVAssetExportPresetMediumQuality];
+    NSString *presetName = [MXSDKOptions sharedInstance].videoConversionPresetName;
+    AVAssetExportSession *exportSession = [AVAssetExportSession exportSessionWithAsset:videoAsset presetName:presetName];
     exportSession.outputURL = outputVideoLocalURL;
+    
+    if (targetFileSize > 0)
+    {
+        // Reduce the target file size by 10% as fileLengthLimit isn't a hard limit
+        exportSession.fileLengthLimit = targetFileSize * 0.9;
+    }
     
     // Check output file types supported by the device
     NSArray *supportedFileTypes = exportSession.supportedFileTypes;